\name{SonVariantsChr21}
\docType{data}
\alias{SonVariantsChr21}
\title{Annotated variants on the chromosome 21, human genome 19, son genome}
\description{
SonVariantsChr21 is a data frame consisting of annotated genomic variants
coming from Complete Genomics whole genome sequencing. Data comes from GIAB project, Ashkenazim Trio, sample HG002 run 1
https://sites.stanford.edu/abms/content/giab-reference-materials-and-data
}
\usage{data(SonVariantsChr21)}
\value{test data}
\format{data frame}
\source{https://sites.stanford.edu/abms/content/giab-reference-materials-and-data}
\keyword{datasets}
